+++ service/src/main/java/com/feidian/service/impl/UserServiceImpl.java
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(Integer id) {
+    public User findById(long id) {
         return userMapper.findById(id);
     }
 
@@ -48,7 +48,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(long id) {
         userMapper.deleteUser(id);
     }
